+++ foursight_core/react/api/react_api_base.py
@@ -222,11 +222,26 @@ class ReactApiBase:
         else:
             return f"{self.foursight_instance_url(request)}/react/{env}/login"
 
+    def is_foursight_cgap(self) -> bool:
+        return app.core.APP_PACKAGE_NAME == "foursight" or app.core.APP_PACKAGE_NAME == "foursight-fourfront"
+
     def is_foursight_fourfront(self) -> bool:
-        return app.core.APP_PACKAGE_NAME == "foursight"
+        return app.core.APP_PACKAGE_NAME == "foursight-cgap"
+
+    def is_foursight_smaht(self) -> bool:
+        return app.core.APP_PACKAGE_NAME == "foursight-smaht"
 
     def get_site_name(self) -> str:
-        return "foursight-cgap" if app.core.APP_PACKAGE_NAME == "foursight-cgap" else "foursight-fourfront"
+        # FYI known site name (APP_PACKAGE_NAME value) are:
+        # - foursight
+        # - foursight-cgap
+        # - foursight-smaht
+        if app.core.APP_PACKAGE_NAME == "foursight":
+            # For React, change foursight to foursight-fourfront,
+            # just to be more explicit and reduce possible confusion.
+            return "foursight-fourfront"
+        else:
+            return app.core.APP_PACKAGE_NAME
 
     def get_default_env(self) -> str:
         return self._envs.get_default_env()